@@ -1,6 +1,7 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.13'
+gem 'rails', '4.0.2'
+gem 'sass-rails', '4.0.1'
 gem 'bootstrap-sass', '~> 2.3.2.1'
 gem 'bcrypt-ruby', '3.0.1'
 gem 'faker', '1.0.1'
@@ -15,7 +16,7 @@ gem 'sucker_punch', '~> 1.0.1'
 gem "rest-client", "~> 1.6.7"
 
 group :development, :test do
-  gem 'sqlite3', '1.3.5'
+  gem 'sqlite3', '1.3.8'
   gem 'rspec-rails', '2.11.0'
   gem 'guard-rspec', '1.2.1'
   gem 'guard-spork', '1.2.0'  
@@ -26,8 +27,7 @@ end
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '3.2.5'
-  gem 'coffee-rails', '3.2.2'
+  gem 'coffee-rails', '4.0.1'
   gem 'uglifier', '1.2.3'
 end
 
